Split OpenSSL backend into separate files (DSA)
